Store original and processed picture to Cloud Storage
When validation succeeds the original and processed pictures are
stored to Google Cloud Storage.

A uuid4 is generated, used as the folder name for storing the pictures.
This id will be used later on a the order id.

The API returns a signed url instead of the base64 encoded image.
The UI is updated to use this signed url.